refactor CLI directory structure to remove circular dependencies
