+++ src/common/dependency-manager/local-manager.ts
@@ -3,9 +3,9 @@ import chalk from 'chalk';
 import fs from 'fs-extra';
 import path from 'path';
 import DependencyManager, { DependencyNode, EnvironmentConfig, EnvironmentConfigBuilder, Refs } from '../../dependency-manager/src';
-import { ComponentConfig } from '../../dependency-manager/src/component-config/base';
-import { ComponentConfigBuilder } from '../../dependency-manager/src/component-config/builder';
 import DependencyGraph from '../../dependency-manager/src/graph';
+import { ComponentConfig } from '../../dependency-manager/src/spec/component/base';
+import { ComponentConfigBuilder } from '../../dependency-manager/src/spec/component/builder';
 import { Dictionary } from '../../dependency-manager/src/utils/dictionary';
 import { flattenValidationErrorsWithLineNumbers, ValidationErrors } from '../../dependency-manager/src/utils/errors';
 import { readIfFile } from '../../dependency-manager/src/utils/files';
