+++ src/dependency-manager/src/graph/node/task.ts
@@ -1,7 +1,7 @@
 import { Type } from 'class-transformer';
 import { DependencyNode, DependencyNodeOptions } from '.';
-import { TaskConfig } from '../../task-config/base';
-import { TaskConfigV1 } from '../../task-config/v1';
+import { TaskConfig } from '../../spec/task/base';
+import { TaskConfigV1 } from '../../spec/task/v1';
 
 export interface TaskNodeOptions {
   ref: string;
