+++ src/dependency-manager/src/manager.ts
@@ -1,17 +1,17 @@
 import { deserialize, plainToClass, serialize } from 'class-transformer';
 import { ValidationError } from 'class-validator';
 import { ServiceNode } from '.';
-import { ComponentConfig } from './component-config/base';
-import { ComponentConfigBuilder } from './component-config/builder';
-import { EnvironmentConfig } from './environment-config/base';
-import { EnvironmentConfigBuilder } from './environment-config/builder';
 import DependencyGraph from './graph';
 import IngressEdge from './graph/edge/ingress';
 import ServiceEdge from './graph/edge/service';
 import GatewayNode from './graph/node/gateway';
 import InterfacesNode from './graph/node/interfaces';
 import { TaskNode } from './graph/node/task';
-import { InterfaceSpec } from './service-config/base';
+import { ComponentConfig } from './spec/component/base';
+import { ComponentConfigBuilder } from './spec/component/builder';
+import { EnvironmentConfig } from './spec/environment/base';
+import { EnvironmentConfigBuilder } from './spec/environment/builder';
+import { InterfaceSpec } from './spec/service/base';
 import { Dictionary } from './utils/dictionary';
 import { flattenValidationErrors, ValidationErrors } from './utils/errors';
 import { escapeJSON, interpolateString, normalizeInterpolation, prefixExpressions, removePrefixForExpressions, replaceBrackets } from './utils/interpolation';
