+++ src/dependency-manager/src/spec/base-spec.ts
