+++ src/dependency-manager/src/spec/component/builder.ts
@@ -4,10 +4,10 @@ import { plainToClass } from 'class-transformer';
 import fs from 'fs-extra';
 import yaml from 'js-yaml';
 import path from 'path';
-import { DeploySpec } from '../common/base';
-import { Dictionary } from '../utils/dictionary';
-import { flattenValidationErrorsWithLineNumbers, ValidationErrors } from '../utils/errors';
-import { insertFileDataFromRefs } from '../utils/files';
+import { Dictionary } from '../../utils/dictionary';
+import { flattenValidationErrorsWithLineNumbers, ValidationErrors } from '../../utils/errors';
+import { insertFileDataFromRefs } from '../../utils/files';
+import { DeploySpec } from '../resource/base';
 import { ComponentConfig } from './base';
 import { ComponentConfigV1 } from './v1';
