+++ src/dependency-manager/src/spec/task/v1.ts
@@ -1,7 +1,7 @@
 import { Type } from 'class-transformer';
 import { IsEmpty, IsInstance, IsOptional, IsString, ValidatorOptions } from 'class-validator';
-import { ResourceConfigV1 } from '../common/v1';
-import { validateDictionary, validateNested } from '../utils/validation';
+import { validateDictionary, validateNested } from '../../utils/validation';
+import { ResourceConfigV1 } from '../resource/v1';
 import { TaskConfig } from './base';
 
 export class TaskConfigV1 extends ResourceConfigV1 implements TaskConfig {
